OPTIONS response body
Today after composer update my API stopped working, stating that CORS is not enabled. But it was until then and nothing really changed. After a lot of hair pulling I found out, that if I added a response body to the OPTIONS response, everything works fine again!

I do not yet know why the body is necessary or what has changed. Please merge my pull request or add the content in your way. I am using stack-cors as a dependency and would have to fork and maintain the other project aswell for that small fix.